Fix for type mismatches that showed up on 64-bit platforms

+++ dynchk/libs/libc/recv.c
@@ -4,9 +4,9 @@
 #include <dlfcn.h>
 #include <stddef.h>
 #undef recv
-static int(*funcptr) (int , void * , size_t , int ) = 0;
+static ssize_t(*funcptr) (int , void * , size_t , int ) = 0;
 
-int recv (int arg0 , void * arg1 , size_t arg2 , int arg3 )
+ssize_t recv (int arg0 , void * arg1 , size_t arg2 , int arg3 )
 {
 	if(!funcptr)
 		funcptr = dlsym(RTLD_NEXT, "recv");
@@ -17,7 +17,7 @@ int recv (int arg0 , void * arg1 , size_t arg2 , int arg3 )
 	return funcptr(arg0, arg1, arg2, arg3);
 }
 
-int lsb_recv (int arg0 , void * arg1 , size_t arg2 , int arg3 )
+ssize_t lsb_recv (int arg0 , void * arg1 , size_t arg2 , int arg3 )
 {
 	if(!funcptr)
 		funcptr = dlsym(RTLD_NEXT, "recv");
