+++ dynchk/libs/libc/sendto.c
@@ -5,9 +5,9 @@
 #include <stddef.h>
 #include <sys/socket.h>
 #undef sendto
-static int(*funcptr) (int , const void * , size_t , int , const struct sockaddr * , socklen_t ) = 0;
+static ssize_t(*funcptr) (int , const void * , size_t , int , const struct sockaddr * , socklen_t ) = 0;
 
-int sendto (int arg0 , const void * arg1 , size_t arg2 , int arg3 , const struct sockaddr * arg4 , socklen_t arg5 )
+ssize_t sendto (int arg0 , const void * arg1 , size_t arg2 , int arg3 , const struct sockaddr * arg4 , socklen_t arg5 )
 {
 	if(!funcptr)
 		funcptr = dlsym(RTLD_NEXT, "sendto");
@@ -20,7 +20,7 @@ int sendto (int arg0 , const void * arg1 , size_t arg2 , int arg3 , const struct
 	return funcptr(arg0, arg1, arg2, arg3, arg4, arg5);
 }
 
-int lsb_sendto (int arg0 , const void * arg1 , size_t arg2 , int arg3 , const struct sockaddr * arg4 , socklen_t arg5 )
+ssize_t lsb_sendto (int arg0 , const void * arg1 , size_t arg2 , int arg3 , const struct sockaddr * arg4 , socklen_t arg5 )
 {
 	if(!funcptr)
 		funcptr = dlsym(RTLD_NEXT, "sendto");
